Remove some unnecessary includes
