+++ CMakeLists.txt
@@ -104,6 +104,7 @@ set(SQLB_HDR
 	src/version.h
 	src/sql/sqlitetypes.h
 	src/sql/Query.h
+	src/sql/ObjectIdentifier.h
 	src/csvparser.h
 	src/sqlite.h
 	src/grammar/sqlite3TokenTypes.hpp
@@ -178,6 +179,7 @@ set(SQLB_SRC
 	src/RowLoader.cpp
 	src/sql/sqlitetypes.cpp
 	src/sql/Query.cpp
+	src/sql/ObjectIdentifier.cpp
 	src/sqltextedit.cpp
 	src/docktextedit.cpp
 	src/csvparser.cpp
