+++ src/ColumnDisplayFormatDialog.h
@@ -2,10 +2,11 @@
 #define COLUMNDISPLAYFORMATDIALOG_H
 
 #include <QDialog>
-#include <QString>
 #include <QMap>
 
-#include "sql/sqlitetypes.h"
+#include "sql/ObjectIdentifier.h"
+
+class QString;
 
 class DBBrowserDB;
 
@@ -18,7 +19,7 @@ class ColumnDisplayFormatDialog : public QDialog
     Q_OBJECT
 
 public:
-    explicit ColumnDisplayFormatDialog(DBBrowserDB& db, const sqlb::ObjectIdentifier& tableName, const QString& colname, QString current_format, QWidget* parent = nullptr);
+    explicit ColumnDisplayFormatDialog(DBBrowserDB& db, const sqlb::ObjectIdentifier& tableName, const QString& colname, const QString& current_format, QWidget* parent = nullptr);
     ~ColumnDisplayFormatDialog() override;
 
     QString selectedDisplayFormat() const;
