+++ src/sqlitedb.h
@@ -1,6 +1,7 @@
 #ifndef SQLITEDB_H
 #define SQLITEDB_H
 
+#include "sql/ObjectIdentifier.h"
 #include "sql/sqlitetypes.h"
 
 #include <condition_variable>
